chore: add logDebug and rename error.ts to logger.ts
